// packages/engine/src/domain/shared/enums.ts
export enum HarborType {
  CoastalNatural = 'Coastal (Natural)',
  CoastalBreakwater = 'Coastal (Breakwater)',
  CoastalTideGate = 'Coastal (Tide Gate)',
  RiverNatural = 'River (Natural)',
  RiverBasin = 'River (Basin)',
  RiverTideGate = 'River (Tide Gate)',
  Lake = 'Lake',
  OpenRoadstead = 'Open Roadstead',
  Canal = 'Canal',
  Lagoon = 'Lagoon',
  Unknown = 'Unknown'
}

export enum HarborSize {
  Small = 'Small',
  Medium = 'Medium',
  Large = 'Large',
  VeryLarge = 'Very Large',
  Unknown = 'Unknown'
}